Refactored: Moved add book form logic to AddNewBookSection component
